third_party/blink/web_tests/svg/dynamic-updates/SVGFECompositeElement-dom-in2-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'in2' attribute of the SVGFECompositeElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS overComposite1.getAttribute('in2') is "SourceGraphic"
PASS overComposite1.getAttribute('in2') is "off1"
PASS successfullyParsed is true

TEST COMPLETE
